+++ CHANGELOG.md
@@ -13,6 +13,12 @@ will not have contiguous patch numbers. Initial major and minor releases will be
 in this file without a patch number. Patch version will be included for bug fixes, but
 may not exactly match a publicly released version.
 
+## [Unreleased]
+
+### Changed
+
+* Use ContractParameterParser.ConvertObject to convert object instances to ContractParameter in NeoTestHarness.Extensions.CreateScript (#20)
+
 ## [3.1.10] - 2021-12-14
 
 ### Changed