+++ CHANGELOG.md
@@ -19,10 +19,12 @@ may not exactly match a publicly released version.
 
 * NeoCsc and NeoExpressBatch MSBuild tasks (plus .targets file)
 * Abstract DotNetToolTask for invoking dotnet tools installed globally or locally
+  * Added ValidateVersion virtual to DotNetToolTask (#25)
 * `ScriptBuilder.EmitContractCall` extension methods
 * `NativeContracts` static class 
   * `NeoToken` and `GasToken` contract hashes
 * `Nep17Token` and `NeoToken` contract interfaces for use with `NeoTestHarness`
+
 
 ### Changed